pyproject.toml
[tool.poetry]
name = "abac-rbac-auth"
version = "0.1.0"
description = "abac-rbac-authentication"
authors = ["Daniil Bairachnyi <b.danny.v@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
sqlalchemy = "^2.0.37"
alembic = "^1.14.1"
pydantic-settings = "^2.7.1"
argon2-cffi = "^23.1.0"
pyjwt = "^2.10.1"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
python_version = "3.10"  # Specify the Python version
